refactor: split out formatter into its work package
Right now this is just a move. We may want to move the formatter impls in to
the relevant packages. For now this will be OK.
