+++ deno.json
@@ -1,5 +1,5 @@
 {
-  "workspace": ["agent", "cli", "tools", "which"],
+  "workspace": ["agent", "cli", "tools", "which", "formatter"],
   "tasks": {
     "compile:cli": "deno compile --unstable-raw-imports --allow-run --allow-read --allow-write --allow-env --allow-net -o bin/fay ./cli/main.ts",
     "preflight:lint": "deno lint --unstable-raw-imports .",
