+++ formatter/deno.json
@@ -1,0 +1,14 @@
+{
+  "name": "@fay/formatter",
+  "version": "0.0.1",
+  "exports": "./mod.ts",
+  "imports": {
+    "@cliffy/ansi": "jsr:@cliffy/ansi@^1.0.0-rc.7",
+    "zod": "npm:zod@^3.25.64"
+  },
+  "lint": {
+    "rules": {
+      "exclude": ["no-slow-types"]
+    }
+  }
+}
